+++ main/managers/views/nfc_view.c
@@ -3681,6 +3681,8 @@ void nfc_view_create(void) {
     lv_obj_t *up_label = lv_label_create(scroll_up_btn);
     lv_label_set_text(up_label, LV_SYMBOL_UP);
     lv_obj_center(up_label);
+    /* hide scroll buttons until we know whether the list is scrollable */
+    lv_obj_add_flag(scroll_up_btn, LV_OBJ_FLAG_HIDDEN);
 
     scroll_down_btn = lv_btn_create(root);
     lv_obj_set_size(scroll_down_btn, SCROLL_BTN_SIZE, SCROLL_BTN_SIZE);
@@ -3693,6 +3695,7 @@ void nfc_view_create(void) {
     lv_obj_t *down_label = lv_label_create(scroll_down_btn);
     lv_label_set_text(down_label, LV_SYMBOL_DOWN);
     lv_obj_center(down_label);
+    lv_obj_add_flag(scroll_down_btn, LV_OBJ_FLAG_HIDDEN);
 
     back_btn = lv_btn_create(root);
     lv_obj_set_size(back_btn, SCROLL_BTN_SIZE + 20, SCROLL_BTN_SIZE);
@@ -3711,6 +3714,20 @@ void nfc_view_create(void) {
     highlight_selected();
 
     display_manager_add_status_bar("NFC");
+#ifdef CONFIG_USE_TOUCHSCREEN
+    /* reveal scroll buttons only if the menu is actually scrollable */
+    if (menu_container && lv_obj_is_valid(menu_container)) {
+        lv_coord_t scroll_bottom = lv_obj_get_scroll_bottom(menu_container);
+        lv_coord_t scroll_top = lv_obj_get_scroll_top(menu_container);
+        if (scroll_bottom > 0 || scroll_top > 0) {
+            if (scroll_up_btn && lv_obj_is_valid(scroll_up_btn)) lv_obj_clear_flag(scroll_up_btn, LV_OBJ_FLAG_HIDDEN);
+            if (scroll_down_btn && lv_obj_is_valid(scroll_down_btn)) lv_obj_clear_flag(scroll_down_btn, LV_OBJ_FLAG_HIDDEN);
+        } else {
+            if (scroll_up_btn && lv_obj_is_valid(scroll_up_btn)) lv_obj_add_flag(scroll_up_btn, LV_OBJ_FLAG_HIDDEN);
+            if (scroll_down_btn && lv_obj_is_valid(scroll_down_btn)) lv_obj_add_flag(scroll_down_btn, LV_OBJ_FLAG_HIDDEN);
+        }
+    }
+#endif
 }
 
 void nfc_view_destroy(void) {