+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Updated mfc dictionary with new additions
 - IR remotes and universals menus now show “No .ir files” placeholder when no IR files are found.
+- Only show touchscreen scroll buttons when the options list is scrollable
 
 ## Revival v1.8